+++ src/common/docman/actions/emptytrash.php
@@ -28,18 +28,24 @@ global $g; //group object
 global $dirid; //id of doc_group
 global $group_id; // id of group
 
-/* you must first delete files before dirs because of database constraints */
-$emptyFile = db_query_params('DELETE FROM doc_data WHERE stateid=$1 and group_id=$2',array('2',$group_id));
-if (!$emptyFile) {
-	exit_error(_('Error'),db_error());
-}
-$emptyDir = db_query_params('DELETE FROM doc_groups WHERE stateid=$1 and group_id=$2',array('2',$group_id));
-if (!$emptyDir) {
-	exit_error(_('Error'),db_error());
-}
+if ( !forge_check_perm ('docman', $group_id, 'approve')) {
+	$feedback= _('Docman Action Denied');
+	Header('Location: '.util_make_url('/docman/?group_id='.$group_id.'&feedback='.urlencode($feedback)));
+	exit;
+} else {
 
-$feedback = _('Emptied Trash successfully');
-Header('Location: '.util_make_url('/docman/?group_id='.$group_id.'&view=admin&feedback='.urlencode($feedback)));
-exit;
+	/* you must first delete files before dirs because of database constraints */
+	$emptyFile = db_query_params('DELETE FROM doc_data WHERE stateid=$1 and group_id=$2',array('2',$group_id));
+	if (!$emptyFile) {
+		exit_error(_('Error'),db_error());
+	}
+	$emptyDir = db_query_params('DELETE FROM doc_groups WHERE stateid=$1 and group_id=$2',array('2',$group_id));
+	if (!$emptyDir) {
+		exit_error(_('Error'),db_error());
+	}
 
+	$feedback = _('Emptied Trash successfully');
+	Header('Location: '.util_make_url('/docman/?group_id='.$group_id.'&view=admin&feedback='.urlencode($feedback)));
+	exit;
+}
 ?>
